+++ src/xwayland/launch_helper.rs
@@ -5,9 +5,12 @@ use std::{
 
 use nix::{
     libc::exit,
-    unistd::{fork, ForkResult},
+    sys::signal,
+    unistd::{fork, ForkResult, Pid},
     Error as NixError, Result as NixResult
 };
+
+use super::xserver::exec_xwayland;
 
 /// A handle to a `fork()`'d child process that is used for starting XWayland
 pub struct LaunchHelper(UnixStream);
@@ -32,6 +35,68 @@ impl LaunchHelper {
                 }
             }
             ForkResult::Parent { child: _ } => Ok(Self(me))
+        }
+    }
+
+    /// Fork a child and call exec_xwayland() in it.
+    pub(crate) fn launch(
+        &self,
+        display: u32,
+        wayland_socket: UnixStream,
+        wm_socket: UnixStream,
+        listen_sockets: &[UnixStream],
+        log: &::slog::Logger,
+    ) -> Result<(Pid, UnixStream), ()> {
+        let (status_child, status_me) = UnixStream::pair().map_err(|_| ())?;
+        match unsafe { fork() } {
+            Ok(ForkResult::Parent { child }) => {
+                // we are the main smithay process
+                Ok((child, status_me))
+            }
+            Ok(ForkResult::Child) => {
+                // we are the first child
+                let mut set = signal::SigSet::empty();
+                set.add(signal::Signal::SIGUSR1);
+                set.add(signal::Signal::SIGCHLD);
+                // we can't handle errors here anyway
+                let _ = signal::sigprocmask(signal::SigmaskHow::SIG_BLOCK, Some(&set), None);
+                match unsafe { fork() } {
+                    Ok(ForkResult::Parent { child }) => {
+                        // When we exit(), we will close() this which wakes up the main process.
+                        let _status_child = status_child;
+                        // we are still the first child
+                        let sig = set.wait();
+                        // Parent will wait for us and know from out
+                        // exit status if XWayland launch was a success or not =)
+                        if let Ok(signal::Signal::SIGCHLD) = sig {
+                            // XWayland has exited before being ready
+                            let _ = ::nix::sys::wait::waitpid(child, None);
+                            unsafe { ::nix::libc::exit(1) };
+                        }
+                        unsafe { ::nix::libc::exit(0) };
+                    }
+                    Ok(ForkResult::Child) => {
+                        // we are the second child, we exec xwayland
+                        match exec_xwayland(display, wayland_socket, wm_socket, listen_sockets) {
+                            Ok(x) => match x {},
+                            Err(e) => {
+                                // well, what can we do ?
+                                error!(log, "exec XWayland failed"; "err" => format!("{:?}", e));
+                                unsafe { ::nix::libc::exit(1) };
+                            }
+                        }
+                    }
+                    Err(e) => {
+                        // well, what can we do ?
+                        error!(log, "XWayland second fork failed"; "err" => format!("{:?}", e));
+                        unsafe { ::nix::libc::exit(1) };
+                    }
+                }
+            }
+            Err(e) => {
+                error!(log, "XWayland first fork failed"; "err" => format!("{:?}", e));
+                Err(())
+            }
         }
     }
 }
